# backend/public/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: "O'Secours"
  description: 'API pour gérer des ressources'
  version: '1.0'
servers:
  -
    url: 'http://localhost:8000/api'
    description: 'API Server'
paths:
...
  /associations:
    get:
      tags:
        - Associations
      summary: 'Get all associations'
      description: 'Returns a list of all associations'
      operationId: getAllAssociations
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Association'
    post:
      tags:
        - Associations
      summary: 'Create a new association'
      description: 'Stores a new association in the database'
      operationId: storeAssociation
      requestBody:
        description: 'Data for the new association'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Association'
      responses:
        '201':
          description: 'Association created successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Association'
  '/associations/{id}':
    get:
      tags:
        - Associations
      summary: 'Get an association by ID'
      description: 'Returns a single association'
      operationId: getAssociationById
      parameters:
        -
          name: id
          in: path
          description: 'ID of the association to return'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Association'
        '404':
          description: 'Association not found'
    put:
      tags:
        - Associations
      summary: 'Update an existing association'
      description: 'Updates data of a specified association'
      operationId: updateAssociation
      parameters:
        -
          name: id
          in: path
          description: 'ID of the association to update'
          required: true
          schema:
            type: string
      requestBody:
        description: 'Data to update the association'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Association'
      responses:
        '200':
          description: 'Association updated successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Association'
        '404':
          description: 'Association not found'
    delete:
      tags:
        - Associations
      summary: 'Delete an association'
      description: 'Soft deletes a specified association from the database'
      operationId: deleteAssociation
      parameters:
        -
          name: id
          in: path
          description: 'ID of the association to delete'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Association deleted successfully'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string }
                  association: { $ref: '#/components/schemas/Association' }
                type: object
        '404':
          description: 'Association not found'
...
components:
  schemas:
...
    Association:
      title: Association
      description: 'Association model'
      properties:
        id:
          description: 'ID of the association'
          type: integer
          example: 1
        name:
          description: 'Name of the association'
          type: string
          example: 'Animal Rescue Association'
        description:
          description: 'Description of the association'
          type: string
          example: 'A non-profit organization dedicated to rescuing animals'
        siret:
          description: 'SIRET number of the association'
          type: string
          maxLength: 14
          minLength: 14
          example: '12345678901234'
        rib:
          description: "RIB (Relevé d'Identité Bancaire) of the association"
          type: string
          maxLength: 34
          example: 'FRXX XXXX XXXX XXXX XXXX XXXX XXX'
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'JWT Authorization header using the Bearer scheme.'
      bearerFormat: JWT
      scheme: bearer